vector_search_concepts_demo/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
openai>=1.12.0
typing-extensions>=4.5.0